Refactor api calls into restApi and wsApi files
